PS: Reorganize buildSrc and add api base url
