+++ buildSrc/src/main/java/com/example/buildsrc/Enviroment.kt
@@ -1,0 +1,20 @@
+package com.example.buildsrc
+
+object Enviroment {
+
+    object Base {
+        const val url = "\"https://gateway.marvel.com/v1/public/\""
+    }
+
+    object Sdk {
+        const val compileSdk = 31
+        const val buildTools = "30.0.3"
+        const val minSdk = 23
+        const val targetSdk = 31
+    }
+
+    object Version {
+        const val versionCode = 1
+        const val versionName = "1.0.0"
+    }
+}
